LayoutTests/accessibility/dynamic-table-caption-expected.txt

This test ensures we maintain the right accessibility text for tables after dynamic caption changes.

	AXTitle: Initial caption
	AXDescription:
	AXHelp:
PASS: text.includes('Initial caption') === true
	AXTitle: Second caption
	AXDescription:
	AXHelp:
PASS: text.includes('Second caption') === true
	AXTitle:
	AXDescription:
	AXHelp:
PASS: !text.includes('caption') === true
	AXTitle: Final caption
	AXDescription:
	AXHelp:
PASS: text.includes('Final caption') === true

PASS successfullyParsed is true

TEST COMPLETE
Final caption
Author	Title
Stephen Hawking	A Brief History of Time
Carl Sagan	Cosmos
